code/uhi_index_analytics/requirements.txt
pyspark
pandas
seaborn
matplotlib
geopandas
shapely
rasterio
rioxarray
pyproj
affine
apache-sedona
PyYAML
ijson
tqdm
cdsapi
google-cloud-storage
google-auth
rtree
scipy
scikit-learn
scikit-image
apache-airflow==2.5.1
apache-airflow-providers-apache-spark
# apache-spark
apache-airflow-providers-google
numpy